backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
google-generativeai
python-dotenv
